requirements.txt
a2wsgi==1.10.10
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.22.1
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.4.0
apache-airflow-core==3.0.6
apache-airflow-providers-common-compat==1.7.1
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.0
apache-airflow-providers-fab==1.1.1
apache-airflow-providers-ftp==3.9.1
apache-airflow-providers-http==4.11.1
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apache-airflow-providers-standard==1.3.0
apache-airflow-task-sdk==1.0.6
apispec==6.6.1
argcomplete==3.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backports.strenum==1.3.1
blinker==1.8.2
cachelib==0.9.0
cadwyn==6.0.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docutils==0.16
email_validator==2.1.1
eval_type_backport==0.3.1
exceptiongroup==1.2.1
Faker==37.12.0
fastapi==0.128.8
fastapi-cli==0.0.22
fastapi-cloud-cli==0.12.0
fastar==0.11.0
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1.20240601
googleapis-common-protos==1.63.1
greenlet==3.2.5
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
httpcore==0.16.3
httptools==0.7.1
httpx==0.23.3
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
libcst==1.8.6
limits==3.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.2.0
msgspec==0.20.0
multidict==6.0.5
numpy==2.0.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
packaging==24.0
pandas==2.3.3
pathlib_abc==0.5.2
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
prison==0.2.1
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.11
pycparser==2.22
pydantic==2.13.1
pydantic-extra-types==2.11.1
pydantic-settings==2.11.0
pydantic_core==2.46.1
Pygments==2.18.0
PyJWT==2.8.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
requests-toolbelt==1.0.0
retryhttp==1.4.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.1
rich-argparse==1.5.0
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.18.1
sentry-sdk==2.58.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
starlette==0.49.3
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
tomli==2.4.1
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.12.1
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
uuid6==2025.0.1
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
anyio==4.4.0
apache-airflow==2.9.2
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.0
apache-airflow-providers-fab==1.1.1
apache-airflow-providers-ftp==3.9.1
apache-airflow-providers-http==4.11.1
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
argcomplete==3.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docutils==0.16
email_validator==2.1.1
exceptiongroup==1.2.1
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1.20240601
googleapis-common-protos==1.63.1
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.2.0
multidict==6.0.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
packaging==24.0
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
prison==0.2.1
protobuf==4.25.3
psutil==5.9.8
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.1
rich-argparse==1.5.0
rpds-py==0.18.1
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
typing_extensions==4.12.1
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==1.26.18
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
